Copy jekyll-theme-isotc211 theme contents
Copy contents of https://github.com/ISO-TC211/jekyll-theme-isotc211
theme.  The only items skipped are _layouts/home.html, and
_includes/_title.html, as they were overriden in Geolexica.org.

This very gem is now another Jekyll theme.  This move is meant to make
adding geolexica pages and layouts easier, as Jekyll is pretty weird
when it comes to path lookup, refusing to find source files correctly
unless they are in theme gems.  This issue can be worked with some
patching, but such additional work is really unwelcome at this stage.

Unfortunately, Jekyll also disallows any theme inheritance or mixing-in.
That means that contents of ISO-TC211/jekyll-theme-isotc211 must be
copied to this gem to make them available, which happens in this very
commit.